Update HARD.conf
On Forestry HARD mode, the MJ energy costs of running the Still, Squeezer, and Fermenter are 4x and the amount of biomass produced by the Fermenter is cut in half, which means producing Ethanol costs roughly 8x as much vs. NORMAL mode.

In this pack, that make Ethanol net energy negative unless you use Mahoe (High sappiness) saplings and are producing the best fruit, and are using exactly the GT Basic Distillery, GT Fluid Extractor, and large Railcraft boilers. Charcoal, however, is very much net energy positive. The end result is Ethanol is nerfed severely compared to other renewable energy sources like charcoal, oil/fuel from bees, oilberries, MOX reactors, etc.,.

To compensate for this, I have tested and am using the config changes above. A simple doubling of the Fermenter biomass output is enough for Ethanol to start to compete with charcoal again. It's still not nearly as efficient as charcoal, but it's not as punishing.